# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FarmEasy API - Indoor farm cycle management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cycles:
    get:
      operationId: listCycles
      tags: [cycles]
      summary: List cycles
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ongoing, history]
            default: ongoing
      responses:
        "200":
          description: List of cycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cycle"
    post:
      operationId: createCycle
      tags: [cycles]
      summary: Create a new cycle (seeding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCycleRequest"
      responses:
        "201":
          description: Cycle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cycle"
  /cycles/{id}:
    get:
      operationId: getCycle
      tags: [cycles]
      summary: Get cycle detail with manual checks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cycle detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cycles/{id}/fertigation:
    post:
      operationId: moveCycleToFertigation
      tags: [cycles]
      summary: Move cycle from germination to fertigation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveFertigationRequest"
      responses:
        "200":
          description: Cycle moved to fertigation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cycle"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cycles/{id}/harvest:
    post:
      operationId: completeCycleHarvest
      tags: [cycles]
      summary: Complete harvest for a cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteHarvestRequest"
      responses:
        "200":
          description: Cycle harvest completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cycle"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cycles/{id}/manual-checks:
    get:
      operationId: listManualChecks
      tags: [cycles]
      summary: List manual checks for a cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of manual checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManualCheck"
    post:
      operationId: createManualCheck
      tags: [cycles]
      summary: Create a manual check for a cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManualCheckRequest"
      responses:
        "201":
          description: Manual check created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualCheck"
...
components:
  schemas:
...
    ManualCheck:
      type: object
      properties:
        id:
          type: integer
        cycleId:
          type: integer
        fullTrays:
          type: integer
        halfTrays:
          type: integer
        isBadTrays:
          type: boolean
        issue:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrls:
          type: array
          items:
            type: string
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, cycleId, fullTrays, halfTrays, isBadTrays, photoUrls, createdAt]
    Cycle:
      type: object
      properties:
        id:
          type: integer
        shortId:
          type: string
        seedLotQrCodes:
          type: array
          items:
            type: string
        seedName:
          type: string
        fullTrays:
          type: integer
        halfTrays:
          type: integer
        seedWeightTray:
          type: number
        growthProfileId:
          type: integer
        growthProfileName:
          type: string
        germinationDays:
          type: integer
        fertigationDays:
          type: integer
        seedingDate:
          type: string
        status:
          type: string
          enum: [germination, fertigation, harvest, completed]
        trayPosition:
          type: ["string", "null"]
        germinationStartedAt:
          type: ["string", "null"]
          format: date-time
        fertigationStartedAt:
          type: ["string", "null"]
          format: date-time
        harvestStartedAt:
          type: ["string", "null"]
          format: date-time
        harvestedQty:
          type: ["number", "null"]
        closedAt:
          type: ["string", "null"]
          format: date-time
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        daysOverdueFertigation:
          type: ["integer", "null"]
        daysOverdueHarvest:
          type: ["integer", "null"]
      required:
        - id
        - shortId
        - seedLotQrCodes
        - seedName
        - fullTrays
        - halfTrays
        - seedWeightTray
        - growthProfileId
        - growthProfileName
        - germinationDays
        - fertigationDays
        - seedingDate
        - status
        - createdAt
    CycleDetail:
      allOf:
        - $ref: "#/components/schemas/Cycle"
        - type: object
          properties:
            manualChecks:
              type: array
              items:
                $ref: "#/components/schemas/ManualCheck"
          required: [manualChecks]
    CreateCycleRequest:
      type: object
      properties:
        seedLotQrCodes:
          type: array
          items:
            type: string
          minItems: 1
        seedName:
          type: string
        fullTrays:
          type: integer
        halfTrays:
          type: integer
        seedWeightTray:
          type: number
        growthProfileId:
          type: integer
        seedingDate:
          type: string
        trayPosition:
          type: string
        humidity:
          type: number
        temperature:
          type: number
        ph:
          type: number
        waterLevel:
          type: number
        nutrientMix:
          type: string
      required:
        - seedLotQrCodes
        - seedName
        - fullTrays
        - halfTrays
        - seedWeightTray
        - growthProfileId
        - seedingDate
        - trayPosition
    MoveFertigationRequest:
      type: object
      properties:
        seedLotQrCode:
          type: string
        humidity:
          type: number
        temperature:
          type: number
        ph:
          type: number
        waterLevel:
          type: number
        nutrientMix:
          type: string
      required: [seedLotQrCode]
    CompleteHarvestRequest:
      type: object
      properties:
        fullTrays:
          type: integer
        halfTrays:
          type: integer
        harvestedQty:
          type: number
        trayQrCode:
          type: string
      required: [fullTrays, halfTrays, harvestedQty, trayQrCode]
    CreateManualCheckRequest:
      type: object
      properties:
        fullTrays:
          type: integer
        halfTrays:
          type: integer
        isBadTrays:
          type: boolean
        issue:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrls:
          type: array
          items:
            type: string
      required: [fullTrays, halfTrays, isBadTrays, photoUrls]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]